cherry pick code of prompt
