+++ component/prompt.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"regexp"
 	"strings"
 	"sync"
 
@@ -375,200 +374,6 @@ func (c *promptComponentImpl) checkPromptRepoPermission(ctx context.Context, req
 		}
 	}
 	return &user, nil
-}
-
-func (c *promptComponentImpl) NewConversation(ctx context.Context, req types.ConversationTitleReq) (*database.PromptConversation, error) {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return nil, errors.New("user does not exist")
-	}
-	conversation := database.PromptConversation{
-		UserID:         user.ID,
-		ConversationID: req.Uuid,
-		Title:          req.Title,
-	}
-
-	err = c.promptConvStore.CreateConversation(ctx, conversation)
-	if err != nil {
-		return nil, fmt.Errorf("new conversation error: %w", err)
-	}
-
-	return &conversation, nil
-}
-
-func (c *promptComponentImpl) ListConversationsByUserID(ctx context.Context, currentUser string) ([]database.PromptConversation, error) {
-	user, err := c.userStore.FindByUsername(ctx, currentUser)
-	if err != nil {
-		return nil, errors.New("user does not exist")
-	}
-	conversations, err := c.promptConvStore.FindConversationsByUserID(ctx, user.ID)
-	if err != nil {
-		return nil, fmt.Errorf("find conversations by user %s error: %w", currentUser, err)
-	}
-	return conversations, nil
-}
-
-func (c *promptComponentImpl) GetConversation(ctx context.Context, req types.ConversationReq) (*database.PromptConversation, error) {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return nil, errors.New("user does not exist")
-	}
-	conversation, err := c.promptConvStore.GetConversationByID(ctx, user.ID, req.Uuid, true)
-	if err != nil {
-		return nil, fmt.Errorf("get conversation by id %s error: %w", req.Uuid, err)
-	}
-	return conversation, nil
-}
-
-func (c *promptComponentImpl) SubmitMessage(ctx context.Context, req types.ConversationReq) (<-chan string, error) {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return nil, errors.New("user does not exist")
-	}
-
-	_, err = c.promptConvStore.GetConversationByID(ctx, user.ID, req.Uuid, false)
-	if err != nil {
-		return nil, fmt.Errorf("invalid conversation by uuid %s error: %w", req.Uuid, err)
-	}
-
-	reqMsg := database.PromptConversationMessage{
-		ConversationID: req.Uuid,
-		Role:           UserRole,
-		Content:        req.Message,
-	}
-	_, err = c.promptConvStore.SaveConversationMessage(ctx, reqMsg)
-	if err != nil {
-		return nil, fmt.Errorf("save user prompt input error: %w", err)
-	}
-
-	llmConfig, err := c.llmConfigStore.GetOptimization(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("get llm config error: %w", err)
-	}
-	slog.Debug("use llm", slog.Any("llmConfig", llmConfig))
-	var headers map[string]string
-	err = json.Unmarshal([]byte(llmConfig.AuthHeader), &headers)
-	if err != nil {
-		return nil, fmt.Errorf("parse llm config header error: %w", err)
-	}
-
-	promptPrefix := ""
-	prefix, err := c.promptPrefixStore.Get(ctx)
-	if err != nil {
-		slog.Warn("fail to find prompt prefix", slog.Any("err", err))
-	} else {
-		chs := isChinese(reqMsg.Content)
-		if chs {
-			promptPrefix = prefix.ZH
-		} else {
-			promptPrefix = prefix.EN
-		}
-	}
-
-	reqData := types.LLMReqBody{
-		Model: llmConfig.ModelName,
-		Messages: []types.LLMMessage{
-			{Role: SystemRole, Content: promptPrefix},
-			{Role: UserRole, Content: reqMsg.Content},
-		},
-		Stream:      true,
-		Temperature: 0.2,
-	}
-	if req.Temperature != nil {
-		reqData.Temperature = *req.Temperature
-	}
-
-	slog.Debug("llm request", slog.Any("reqData", reqData))
-	ch, err := c.llmClient.Chat(ctx, llmConfig.ApiEndpoint, headers, reqData)
-	if err != nil {
-		return nil, fmt.Errorf("call llm error: %w", err)
-	}
-	return ch, nil
-}
-
-func (c *promptComponentImpl) SaveGeneratedText(ctx context.Context, req types.Conversation) (*database.PromptConversationMessage, error) {
-	respMsg := database.PromptConversationMessage{
-		ConversationID: req.Uuid,
-		Role:           AssistantRole,
-		Content:        req.Message,
-	}
-	msg, err := c.promptConvStore.SaveConversationMessage(ctx, respMsg)
-	if err != nil {
-		return nil, fmt.Errorf("save system generated response error: %w", err)
-	}
-	return msg, nil
-}
-
-func (c *promptComponentImpl) RemoveConversation(ctx context.Context, req types.ConversationReq) error {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return errors.New("user does not exist")
-	}
-
-	err = c.promptConvStore.DeleteConversationsByID(ctx, user.ID, req.Uuid)
-	if err != nil {
-		return fmt.Errorf("remove conversation error: %w", err)
-	}
-	return nil
-}
-
-func (c *promptComponentImpl) UpdateConversation(ctx context.Context, req types.ConversationTitleReq) (*database.PromptConversation, error) {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return nil, errors.New("user does not exist")
-	}
-
-	err = c.promptConvStore.UpdateConversation(ctx, database.PromptConversation{
-		UserID:         user.ID,
-		ConversationID: req.Uuid,
-		Title:          req.Title,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("update conversation title error: %w", err)
-	}
-
-	resp, err := c.promptConvStore.GetConversationByID(ctx, user.ID, req.Uuid, false)
-	if err != nil {
-		return nil, fmt.Errorf("invalid conversation by uuid %s error: %w", req.Uuid, err)
-	}
-	return resp, nil
-}
-
-func (c *promptComponentImpl) LikeConversationMessage(ctx context.Context, req types.ConversationMessageReq) error {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return errors.New("user does not exist")
-	}
-	_, err = c.promptConvStore.GetConversationByID(ctx, user.ID, req.Uuid, false)
-	if err != nil {
-		return fmt.Errorf("invalid conversation by uuid %s error: %w", req.Uuid, err)
-	}
-	err = c.promptConvStore.LikeMessageByID(ctx, req.Id)
-	if err != nil {
-		return fmt.Errorf("update like message by id %d error: %w", req.Id, err)
-	}
-	return nil
-}
-
-func (c *promptComponentImpl) HateConversationMessage(ctx context.Context, req types.ConversationMessageReq) error {
-	user, err := c.userStore.FindByUsername(ctx, req.CurrentUser)
-	if err != nil {
-		return errors.New("user does not exist")
-	}
-	_, err = c.promptConvStore.GetConversationByID(ctx, user.ID, req.Uuid, false)
-	if err != nil {
-		return fmt.Errorf("invalid conversation by uuid %s error: %w", req.Uuid, err)
-	}
-	err = c.promptConvStore.HateMessageByID(ctx, req.Id)
-	if err != nil {
-		return fmt.Errorf("update hate message by id %d error: %w", req.Id, err)
-	}
-	return nil
-}
-
-func isChinese(s string) bool {
-	re := regexp.MustCompile(`[\p{Han}]`)
-	return re.MatchString(s)
 }
 
 func (c *promptComponentImpl) SetRelationModels(ctx context.Context, req types.RelationModels) error {
