+++ component/prompt_test.go
@@ -3,15 +3,10 @@ package component
 import (
 	"context"
 	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"testing"
 
-	"github.com/alibabacloud-go/tea/tea"
-	"github.com/jarcoal/httpmock"
-	"github.com/spf13/cast"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"opencsg.com/csghub-server/builder/git/gitserver"
@@ -359,231 +354,6 @@ func TestPromptComponent_GetPrompt(t *testing.T) {
 	require.Nil(t, err)
 	require.Equal(t, "ai", output.Title)
 	require.Equal(t, "foo.jsonl", output.FilePath)
-}
-
-func TestPromptComponent_NewConversation(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().CreateConversation(ctx, database.PromptConversation{
-		UserID:         123,
-		ConversationID: "zzz",
-		Title:          "test",
-	}).Return(nil).Once()
-
-	cv, err := pc.NewConversation(ctx, types.ConversationTitleReq{
-		CurrentUser: "foo",
-		ConversationTitle: types.ConversationTitle{
-			Uuid:  "zzz",
-			Title: "test",
-		},
-	})
-	require.Nil(t, err)
-	require.Equal(t, 123, int(cv.UserID))
-
-}
-
-func TestPromptComponent_ListConversationByUserID(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	mockedResults := []database.PromptConversation{
-		{Title: "foo"},
-		{Title: "bar"},
-	}
-	pc.mocks.stores.PromptConversationMock().EXPECT().FindConversationsByUserID(ctx, int64(123)).Return(mockedResults, nil).Once()
-
-	results, err := pc.ListConversationsByUserID(ctx, "foo")
-	require.Nil(t, err)
-	require.Equal(t, mockedResults, results)
-
-}
-
-func TestPromptComponent_GetConversation(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	mocked := &database.PromptConversation{}
-	pc.mocks.stores.PromptConversationMock().EXPECT().GetConversationByID(ctx, int64(123), "uuid", true).Return(mocked, nil).Once()
-
-	cv, err := pc.GetConversation(ctx, types.ConversationReq{
-		CurrentUser: "foo",
-		Conversation: types.Conversation{
-			Uuid: "uuid",
-		},
-	})
-	require.Nil(t, err)
-	require.Equal(t, mocked, cv)
-
-}
-
-func TestPromptComponent_SubmitMessage(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	for _, lang := range []string{"en", "zh"} {
-		t.Run(lang, func(t *testing.T) {
-
-			pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-			pc.mocks.stores.PromptConversationMock().EXPECT().GetConversationByID(ctx, int64(123), "uuid", false).Return(&database.PromptConversation{}, nil).Once()
-
-			content := "go"
-			if lang == "zh" {
-				content = "围棋"
-			}
-			pc.mocks.stores.PromptConversationMock().EXPECT().SaveConversationMessage(
-				ctx, database.PromptConversationMessage{
-					ConversationID: "uuid",
-					Role:           UserRole,
-					Content:        content,
-				},
-			).Return(&database.PromptConversationMessage{}, nil)
-			pc.mocks.stores.LLMConfigMock().EXPECT().GetOptimization(ctx).Return(&database.LLMConfig{
-				ApiEndpoint: "https://llm.com",
-				AuthHeader:  `{"token": "foobar"}`,
-			}, nil).Once()
-			pc.mocks.stores.PromptPrefixMock().EXPECT().Get(ctx).Return(&database.PromptPrefix{
-				ZH: "use Chinese",
-				EN: "use English",
-			}, nil).Once()
-			httpmock.Activate()
-			t.Cleanup(httpmock.DeactivateAndReset)
-
-			httpmock.RegisterResponder("POST", "https://llm.com",
-				func(req *http.Request) (*http.Response, error) {
-					article := make(map[string]interface{})
-					if err := json.NewDecoder(req.Body).Decode(&article); err != nil {
-						return httpmock.NewStringResponse(400, ""), nil
-					}
-					prefix := cast.ToStringMap(cast.ToSlice(article["messages"])[0])["content"]
-					d := ""
-					switch prefix {
-					case "use English":
-						d = `[{"id": 1, "name": "My Great Article"}]`
-					case "use Chinese":
-						d = `[{"id": 1, "name": "好好好"}]`
-					default:
-						d = "wrong"
-					}
-					return httpmock.NewStringResponse(
-						200, d,
-					), nil
-				})
-
-			ch, err := pc.SubmitMessage(ctx, types.ConversationReq{
-				CurrentUser: "foo",
-				Conversation: types.Conversation{
-					Uuid:    "uuid",
-					Message: content,
-				},
-			})
-			require.Nil(t, err)
-			all := ""
-			for i := range ch {
-				all += i
-			}
-			if lang == "en" {
-				require.Equal(t, "[{\"id\": 1, \"name\": \"My Great Article\"}]", all)
-			} else {
-				require.Equal(t, "[{\"id\": 1, \"name\": \"好好好\"}]", all)
-			}
-		})
-	}
-}
-
-func TestPromptComponent_SaveGeneratedText(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	mocked := &database.PromptConversationMessage{}
-	pc.mocks.stores.PromptConversationMock().EXPECT().SaveConversationMessage(ctx, database.PromptConversationMessage{
-		ConversationID: "uuid",
-		Role:           AssistantRole,
-		Content:        "m",
-	}).Return(mocked, nil).Once()
-
-	m, err := pc.SaveGeneratedText(ctx, types.Conversation{
-		Uuid:        "uuid",
-		Message:     "m",
-		Temperature: tea.Float64(0.8),
-	})
-	require.Nil(t, err)
-	require.Equal(t, mocked, m)
-}
-
-func TestPromptComponent_RemoveConversation(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().DeleteConversationsByID(ctx, int64(123), "uuid").Return(nil).Once()
-
-	err := pc.RemoveConversation(ctx, types.ConversationReq{
-		CurrentUser: "foo",
-		Conversation: types.Conversation{
-			Uuid: "uuid",
-		},
-	})
-	require.Nil(t, err)
-}
-
-func TestPromptComponent_UpdateConversation(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().UpdateConversation(ctx, database.PromptConversation{
-		UserID:         123,
-		ConversationID: "uuid",
-		Title:          "title",
-	}).Return(nil).Once()
-	mocked := &database.PromptConversation{}
-	pc.mocks.stores.PromptConversationMock().EXPECT().GetConversationByID(ctx, int64(123), "uuid", false).Return(mocked, nil)
-
-	cv, err := pc.UpdateConversation(ctx, types.ConversationTitleReq{
-		CurrentUser: "foo",
-		ConversationTitle: types.ConversationTitle{
-			Uuid:  "uuid",
-			Title: "title",
-		},
-	})
-	require.Nil(t, err)
-	require.Equal(t, mocked, cv)
-}
-
-func TestPromptComponent_LikeConversationMessage(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().GetConversationByID(ctx, int64(123), "uuid", false).Return(&database.PromptConversation{}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().LikeMessageByID(ctx, int64(123)).Return(nil).Once()
-
-	err := pc.LikeConversationMessage(ctx, types.ConversationMessageReq{
-		Uuid:        "uuid",
-		Id:          123,
-		CurrentUser: "foo",
-	})
-	require.Nil(t, err)
-}
-
-func TestPromptComponent_HateConversationMessage(t *testing.T) {
-	ctx := context.TODO()
-	pc := initializeTestPromptComponent(ctx, t)
-
-	pc.mocks.stores.UserMock().EXPECT().FindByUsername(ctx, "foo").Return(database.User{ID: 123}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().GetConversationByID(ctx, int64(123), "uuid", false).Return(&database.PromptConversation{}, nil).Once()
-	pc.mocks.stores.PromptConversationMock().EXPECT().HateMessageByID(ctx, int64(123)).Return(nil).Once()
-
-	err := pc.HateConversationMessage(ctx, types.ConversationMessageReq{
-		Uuid:        "uuid",
-		Id:          123,
-		CurrentUser: "foo",
-	})
-	require.Nil(t, err)
 }
 
 func TestPromptComponent_SetRelationModels(t *testing.T) {
